Finishing up pseudo-loglikelihood calculations.

+++ recogners/core/dataset.py
@@ -56,7 +56,7 @@ class Dataset(data.Dataset):
         """
 
         # Gets a sample based on its index
-        x = torch.from_numpy(self.X[index]).float()
+        x = torch.from_numpy(self.X[index]).double()
 
         # Gets a sample's label based on its index
         y = torch.from_numpy((self.Y[index]).reshape([1, 1]))
